Fix stylecop warning
